fix: ts errors
